+ Upload Optimization based on last modified date
+ Adding code to optimize sync/upload operation .

+ A dictionary keeps track of files and last modified dates , this is
written to the file (of name last_modified.txt)

+ When user syncs the folder, cloud crate compares the current
last_modified date of the file with the previous last modified date
from the dictionary ; uploads if the file had changed , skips uploading
if no change
=============================================
Workflow:
——————
1. It might be the first time a user is syncing a folder - in that case
last_modified.txt file is missing.
2. last_modified.txt folder is constructed and all files are uploaded.

3. It could be a case , where user is re-syncing . In this scenario ,
last_modified.txt file is already available .
4. Only files that have update last_modfied timestamps are uploaded.

Known bug : if there is just one new file and no changes to existing
files, the last_modified.txt file is present but entry for the new file
is missing. we have to add this file to the dictionary and upload . but
the script currently failover to the catch block and uploads all the
files.