Fixes the following Known bug from the previous commit: ============== Known bug : if there is just one new file and no changes to existing files, the last_modified.txt file is present but entry for the new file is missing. we have to add this file to the dictionary and upload . but the script currently fails over to the catch block and uploads all the files. ========= The catch block was catching two errors : 1. One was IOError (i.e.) the last_modified.txt file is not present , it indicates to the application that its the first ever time , the user is syncing and we build our house keeping dictionary
2. It was also catching a keyError (i.e.) the last_modified.txt file is
present , BUT the dictionary it holds doesn’t have a key,value entry
for the file that was added after the last sync.
It was failing over and catch block was uploading all the files.

The fix uploads only the new added file instead of all files in case of
KeyError in the book keeping dictionary.